% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_markers.R
\name{add_markers}
\alias{add_markers}
\title{Add markers}
\usage{
add_markers(
  marks = NULL,
  lifemap,
  radius = 10,
  popup = "none",
  form = "none",
  cluster = "none"
)
}
\arguments{
\item{marks}{data.frame of marks to view on the lifemap, given by the user;
NULL by default}

\item{lifemap}{the map generated}

\item{radius}{the size of markers}

\item{popup}{if user wants popups ; "none" by default}

\item{form}{if user wants special form ; "none" by default}

\item{cluster}{if user wants cluster ; "none" by default}
}
\value{
mark_lifemap
}
\description{
Add of markers on the lifemap, with various options for form, cluster and popup
}
\examples{
add_markers(marks, newmap(solr_request(2, "ncbi"), "ncbi"))

}
