% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_minimap.R
\name{add_minimap}
\alias{add_minimap}
\title{add_minimap
adds a minimap}
\usage{
add_minimap(lifemap, map = "standard")
}
\arguments{
\item{lifemap}{a lifemap object}

\item{map}{a character string for a lifemap, "standard" by default}
}
\value{
withMiniMap
}
\description{
add_minimap
adds a minimap
}
\examples{
add_minimap(newmap())

}
