% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ascend.R
\name{get_ascend}
\alias{get_ascend}
\title{Get ascend}
\usage{
get_ascend(taxids)
}
\arguments{
\item{taxids}{a vector of taxids}
}
\value{
ascend a vector of taxids
}
\description{
Get a vector of taxids from maximum 1000 of the ascendants from each taxid
in input.
}
\examples{
get_ascend(2)

}
