% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_descend.R
\name{get_descend}
\alias{get_descend}
\title{get_descend}
\usage{
get_descend(taxids)
}
\arguments{
\item{taxids}{a vector of taxids}
}
\value{
DATA a vector of taxids
}
\description{
Get a vector of taxids from maximum 1000 of the descendant from each taxid in input
}
\examples{
get_descend(2)

}
