% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmap.R
\name{newmap}
\alias{newmap}
\title{Newmap}
\usage{
newmap(coordinates = NULL, map = "standard")
}
\arguments{
\item{coordinates}{of groups, as a dataframe with a \emph{coordinate} column being the result of Solr request; NULL by default}

\item{map}{a character string for the version of lifemap among 'standard', 'ncbi', 'french' and virus' version; 'standard' by default}
}
\value{
A new map widget used by \code{leaflet} and \code{marklifemap}.
}
\description{
Create a new lifemap according to the chosen \code{map}.
Requires a connection to \url{http://lifemap.univ-lyon1.fr/} and its database.
}
\examples{
newmap(map='virus')

}
