% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solr_request.R
\name{solr_request}
\alias{solr_request}
\title{Solr request}
\usage{
solr_request(taxids, map = "standard")
}
\arguments{
\item{taxids}{a vector of taxids}

\item{map}{a character string for a lifemap, "standard" by default}
}
\value{
coordinates
}
\description{
Solr request
}
\examples{
solr_request(c(2,9443,2087), map='ncbi')
}
