% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_minimap.R
\name{add_minimap}
\alias{add_minimap}
\title{Add minimap}
\usage{
add_minimap(lifemap, map = "standard")
}
\arguments{
\item{lifemap}{a lifemap object}

\item{map}{a character string for a lifemap, "standard" by default}
}
\value{
with_minimap the lifemap with the minimap
}
\description{
Function to add a little map at the right top of the window, to see
where you are navigating on the lifemap.
}
\examples{
add_minimap(newmap())

}
