% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marklifemap.R
\name{marklifemap}
\alias{marklifemap}
\title{Create markers on a Lifemap.}
\usage{
marklifemap(
  data,
  type = "taxid",
  map = "standard",
  minimap = FALSE,
  ascendant = FALSE,
  descendant = FALSE,
  popup = "none",
  form = "none",
  cluster = "none",
  output = FALSE,
  filename = "result.csv"
)
}
\arguments{
\item{data}{a dataframe or vector of information}

\item{type}{a character string to choose the type of data used for \code{solr_request} ; must be identical to an attribute from the SolR database, and at least one column name if \code{data} is a dataframe}

\item{map}{a character string : basemap among 'standard', virus', 'ncbi' and 'french' versions of lifemap}

\item{minimap}{a boolean to choose if the minimap should be displayed}

\item{ascendant}{a boolean to choose the display of ascendant groups}

\item{descendant}{a boolean to choose the display of descendant groups}

\item{popup}{a character string to choose the type of popup among 'dataframe', 'sci_info' and 'none'; 'none' by default}

\item{form}{a character string to choose the form of markers among 'dataframe', 'nbdesc' and 'none'; 'none' by default}

\item{cluster}{a character sting to choose the type of cluster among 'sum' and 'none'; 'none' by default}

\item{output}{a character string to choose if search result should be displayed among 'file', 'merge_file', 'console' and 'merge_console'; no output by default}

\item{filename}{a character string to name the output file if 'file' is chosen for \code{output} ; written in working directory}
}
\value{
the lifemap widget with markers
}
\description{
Create a Lifemap using \code{leaflet()} and add markers from \code{data} depending on the options choosed by user.
Requires a connexion to \url{http://lifemap.univ-lyon1.fr/} and its database.
}
\examples{
marklifemap(data.frame(list("taxid"=c(2,9443,2087),"name"=c('A','B','C'),"nbdesc"=c(300,20,109))), map='ncbi')

}
