% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newmap.R
\name{newmap}
\alias{newmap}
\title{Newmap}
\usage{
newmap(coordinates = NULL, map = "standard")
}
\arguments{
\item{coordinates}{of taxids, the result of Solr request; NULL by default}

\item{map}{the version of lifemap among 'standard', 'ncbi', 'french' and
'virus' version; 'standard' by default}
}
\value{
lifemap
}
\description{
Function to create a new lifemap according to the chosen map.
Keeps coordinates of taxons in memory if there are.
}
\examples{
newmap(map='virus')

}
