% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solr_request.R
\name{solr_request}
\alias{solr_request}
\title{Solr request}
\usage{
solr_request(info, type = "taxid", map = "standard", input = TRUE)
}
\arguments{
\item{info}{a vector of information}

\item{type}{a character string for the type of data used}

\item{map}{a character string for a lifemap, "standard" by default}

\item{input}{a boolean to avoid overwriting "unfound.txt"}
}
\value{
A dataframe with information from the query extracted. Can be used for a leaflet, lifemap or marklifemap map.
}
\description{
Search in the \url{http://lifemap.univ-lyon1.fr/} database for \code{info} in the \code{type} column.
Call \code{map_and_data} to write all elements of \code{info} not found in the database in an \code{unfound.txt} file in working directory.
}
\examples{
solr_request(c(2,9443,2087), map='ncbi')
}
